+++ test/warpath/query/index_operator_test.exs
@@ -1,4 +1,4 @@
-defmodule Warpath.Query.ArrayIndexOperatorTest do
+defmodule Warpath.Query.IndexOperatorTest do
   use ExUnit.Case, async: true
   use ExUnitProperties
 
@@ -10,20 +10,20 @@ defmodule Warpath.Query.ArrayIndexOperatorTest do
 
   @relative_path [{:root, "$"}]
 
-  defp env_for_array_index(indexes, previous_operator \\ nil) do
+  defp env_for_indexes(indexes, previous_operator \\ nil) do
     access = Enum.map(indexes, &{:index_access, &1})
     Env.new({:indexes, access}, previous_operator)
   end
 
   test "evaluate a empty list always result in an element with empty list" do
-    IndexOperator.evaluate([], @relative_path, env_for_array_index([1, 2, 3]))
+    IndexOperator.evaluate([], @relative_path, env_for_indexes([1, 2, 3]))
   end
 
   property "always unwrap element when evaluate a single existent index" do
     check all terms <- list_of(term(), min_length: 1),
               index = Enum.random(Range.new(0, length(terms) - 1)),
               item = Enum.at(terms, index) do
-      result = IndexOperator.evaluate(terms, @relative_path, env_for_array_index([index]))
+      result = IndexOperator.evaluate(terms, @relative_path, env_for_indexes([index]))
       assert result == Element.new(item, [{:index_access, index} | @relative_path])
     end
   end
@@ -31,7 +31,7 @@ defmodule Warpath.Query.ArrayIndexOperatorTest do
   property "postive out of bound index are never evaluated" do
     check all terms <- list_of(term(), min_length: 1),
               count = length(terms) do
-      env = env_for_array_index([count, count + 1])
+      env = env_for_indexes([count, count + 1])
       assert IndexOperator.evaluate(terms, @relative_path, env) == []
     end
   end
@@ -39,14 +39,14 @@ defmodule Warpath.Query.ArrayIndexOperatorTest do
   property "negative out of bound index are never evaluated" do
     check all terms <- list_of(term(), min_length: 1),
               count = length(terms) do
-      env = env_for_array_index([-(count + 1), -(count + 2)])
+      env = env_for_indexes([-(count + 1), -(count + 2)])
       assert IndexOperator.evaluate(terms, @relative_path, env) == []
     end
   end
 
   property "flatten it's result on traverse a list of element after wildcard operator" do
     element = map(list_of(term(), length: 2), &Element.new(&1, []))
-    env = env_for_array_index([0, 1])
+    env = env_for_indexes([0, 1])
 
     check all elements <- list_of(element) do
       result = IndexOperator.evaluate(elements, [], env)
@@ -66,7 +66,7 @@ defmodule Warpath.Query.ArrayIndexOperatorTest do
   end
 
   property "index query are only evaluated on list data type" do
-    env = env_for_array_index([0])
+    env = env_for_indexes([0])
 
     check all term <- term(),
               element = Element.new(term, []) do
@@ -76,7 +76,7 @@ defmodule Warpath.Query.ArrayIndexOperatorTest do
   end
 
   property "evaluate multipe index" do
-    env = env_for_array_index([0, 1])
+    env = env_for_indexes([0, 1])
 
     check all terms <- list_of(term(), length: 2) do
       result = IndexOperator.evaluate(terms, [], env)
@@ -93,7 +93,7 @@ defmodule Warpath.Query.ArrayIndexOperatorTest do
   property "can evaluate negative index" do
     check all terms <- list_of(term(), length: 2),
               count = length(terms) do
-      env = env_for_array_index([-count, -(count - 1)])
+      env = env_for_indexes([-count, -(count - 1)])
       result = IndexOperator.evaluate(terms, [], env)
 
       [first, second | _] = terms
@@ -103,5 +103,9 @@ defmodule Warpath.Query.ArrayIndexOperatorTest do
                Element.new(second, [{:index_access, 1}])
              ]
     end
+  end
+
+  test "evaluate/3 is nil safe" do
+    assert IndexOperator.evaluate(nil, [], env_for_indexes([1, 2, 3])) == []
   end
 end